+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Dropped support for .NET 6
 - Dropped support for .NET 7
 - Added support for .NET 8
+- Added `MergeDefaultClassTagHelper` which will merge the `default-class` attribute with the `class` attribute
 
 ## [0.5.0](https://github.com/xt0rted/tailwindcss-tag-helpers/compare/v0.4.0...v0.5.0) - 2023-02-24